refactor(bundler-vite): extract public logic to `@vuepress/vite-kit`
