+++ packages/bundler-vite/src/plugins/index.ts
@@ -1,3 +1,1 @@
-export * from './vuepressMainPlugin.js'
 export * from './vuepressUserConfigPlugin.js'
-export * from './vuepressVuePlugin.js'
