+++ packages/bundler-vite/src/viteBundler.ts
@@ -1,5 +1,5 @@
 import type { Bundler } from '@vuepress/core'
-import { build } from './build/index.js'
+import { build } from './build.js'
 import { dev } from './dev.js'
 import type { ViteBundlerOptions } from './types.js'
