+++ packages/vite-kit/client.d.ts
