+++ packages/vite-kit/package.json
@@ -1,11 +1,9 @@
 {
-  "name": "@vuepress/bundler-vite",
+  "name": "@vuepress/vite-kit",
   "version": "2.0.0-rc.14",
-  "description": "Bundler vite package of VuePress",
+  "description": "Utils for vite-compatible VuePress bundlers",
   "keywords": [
-    "vuepress-bundler",
     "vuepress",
-    "bundler",
     "vite"
   ],
   "homepage": "https://github.com/vuepress",
@@ -35,15 +33,19 @@
     "clean": "rimraf dist"
   },
   "dependencies": {
+    "@vitejs/plugin-vue": "^5.0.5",
     "@vuepress/client": "workspace:*",
     "@vuepress/core": "workspace:*",
+    "@vuepress/shared": "workspace:*",
     "@vuepress/utils": "workspace:*",
-    "@vuepress/vite-kit": "workspace:*",
+    "autoprefixer": "^10.4.19",
+    "connect-history-api-fallback": "^2.0.0",
+    "postcss": "^8.4.38",
+    "postcss-load-config": "^6.0.1",
+    "rollup": "^4.18.0",
     "vite": "~5.3.1",
-    "vue": "^3.4.29"
-  },
-  "devDependencies": {
-    "rollup": "^4.18.0"
+    "vue": "^3.4.29",
+    "vue-router": "^4.3.3"
   },
   "publishConfig": {
     "access": "public"
