+++ packages/vite-kit/src/index.ts
@@ -1,0 +1,3 @@
+export type { Options as VuePluginOptions } from '@vitejs/plugin-vue'
+export * from './build/index.js'
+export * from './plugins/index.js'
