+++ packages/vite-kit/src/plugins/index.ts
@@ -1,0 +1,2 @@
+export * from './vuepressMainPlugin.js'
+export * from './vuepressVuePlugin.js'
