+++ packages/vite-kit/src/plugins/vuepressMainPlugin.ts
