+++ packages/vite-kit/src/types.ts
@@ -1,0 +1,1 @@
+export type { Options as VuePluginOptions } from '@vitejs/plugin-vue'
